SEQNG-270 Refactor tests
